USERS WORKS NOWWWWW

+++ routes/todoRouter.js
@@ -1,13 +1,16 @@
 const express = require('express');
-const Todo = require('../models/todo');
+const { Todo } = require('../models/todo');
 const authenticate = require('../authenticate');
+const cors = require('./cors');
 
 const todoRouter = express.Router();
 
 todoRouter
     .route('/')
-    .get(authenticate.verifyUser, (req, res, next) => {
-        Todo.find()
+    .options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+
+    .get(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+        Todo.find({ userId: req.user._id })
             .then((todos) => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -16,24 +19,25 @@ todoRouter
             .catch((err) => next(err));
     })
 
-    .post(authenticate.verifyUser, (req, res, next) => {
-        Todo.create(req.body)
-            .then((todo) => {
-                console.log('Todo created', todo);
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(todo);
+    .post(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+        const newTodo = new Todo({ ...req.body, userId: req.user._id });
+        newTodo.save()
+            .then(savedTodo => {
+                res.status(201).json(savedTodo);
             })
-            .catch((err) => next(err));
+            .catch(err => {
+                next(err);
+            });
     })
+    
 
-    .put(authenticate.verifyUser, (req, res) => {
+    .put(cors.corsWithOptions, authenticate.verifyUser, (req, res) => {
         res.statusCode = 403;
         res.end('PUT opration not supported on /todos');
     })
 
-    .delete(authenticate.verifyUser, (req, res, next) => {
-        Todo.deleteMany()
+    .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+        Todo.deleteMany({ userId: req.user._id })
             .then((response) => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -44,7 +48,8 @@ todoRouter
 
 todoRouter
     .route('/:todoId')
-    .get(authenticate.verifyUser, (req, res, next) => {
+    .options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+    .get(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         Todo.findById(req.params.todoId)
             .then((todo) => {
                 res.statusCode = 200;
@@ -54,12 +59,12 @@ todoRouter
             .catch((err) => next(err));
     })
 
-    .post(authenticate.verifyUser, (req, res) => {
+    .post(cors.corsWithOptions, authenticate.verifyUser, (req, res) => {
         res.status = 403;
         res.end(`POST operation not supported on /todos/${req.params.todoId}`);
     })
 
-    .put(authenticate.verifyUser, (req, res, next) => {
+    .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         Todo.findByIdAndUpdate(
             req.params.todoId,
             { $set: req.body },
@@ -73,7 +78,7 @@ todoRouter
             .catch((err) => next(err));
     })
 
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
         Todo.findByIdAndDelete(req.params.todoId)
             .then((response) => {
                 res.statusCode = 200;
